+++ release-notes.md
@@ -2,6 +2,7 @@ Version 1.1.2 (unreleased)
 --------------
 * Added Vec4i.toVec3d()
 * Added Vec3d.set(Vector3f)
+* Added Quatd.set(Quaternion)
 
 Version 1.1.1 
 --------------